requirements.txt

Django==5.0.7
django-ckeditor==6.7.1
pillow==10.4.0
gunicorn==20.0.4
whitenoise==5.2.0